zero/requirements.txt
blinker==1.7.0
click==8.1.7
colorzero==2.0
flask==3.0.1
gpiozero==2.0
lgpio=0.2.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pytz==2023.3.post1
RPi.bme280==0.2.4
smbus2==0.4.3
werkzeug==3.0.1
zipp==3.17.0
